+++ tests/garage-named-up/devenv.nix
@@ -1,0 +1,8 @@
+{ ... }:
+{
+  services.garage = {
+    enable = true;
+    rpcSecret = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa";
+    buckets = [ "named-up-bucket" ];
+  };
+}